requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
supabase==2.3.0
httpx>=0.24.0,<0.25.0  # Adjusted to match supabase requirement
pydantic==2.5.2
pydantic-settings==2.1.0  # Added for config settings
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pytest==7.4.3
python-dotenv==1.0.0
langchain==0.1.1
pinecone-client==3.0.0
pytest-asyncio==0.21.1
asyncpg==0.29.0
bcrypt==4.0.1
